refactor: remove normal dispatch wrapper
