+++ crates/evm2/src/interpreter/instructions/table/normal_packed.rs
@@ -19,7 +19,7 @@ pub(super) type RawInstrFn<T> = extern_table!(
     ) -> InstrFnRet
 );
 
-super::normal::normal_tables!();
+normal_tables!();
 
 pub(crate) type LoopState = RemainingGas;
